verus_compare helpers to lynette lib
